pyproject.toml
[tool.poetry]
name = "express24"
version = "0.1.0"
description = ""
authors = ["nmunisa06 <munisa06nm@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
django = "5.1.2"
django-jazzmin = "3.0.1"
psycopg2-binary = "2.9.10"
python-dotenv = "1.0.1"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
djangorestframework = "^3.15.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
